// src/components/rich-text/rich-text-options.interface.ts
export interface RichTextOptions {
    undo?: boolean;
    redo?: boolean;
    bold?: boolean;
    italic?: boolean;
    underline?: boolean;
    largeText?: boolean;
    smallText?: boolean;
    alignLeft?: boolean;
    alignCenter?: boolean;
    alignRight?: boolean;
    justify?: boolean;
    lineJump?: boolean;
    orderedList?: boolean;
    unorderedList?: boolean;
    strikethrough?: boolean;
    canClose?: boolean;
  }
  